* origin/released:
  fontconfig libX11 libXmu libxcb mesa xserver git update 10 Sep 2013

Conflicts:
	mesalib/src/glsl/.gitignore
	xorg-server/hw/xwin/InitOutput.c
	xorg-server/hw/xwin/winclipboardwndproc.c
	xorg-server/hw/xwin/winmultiwindowwm.c
	xorg-server/hw/xwin/winsetsp.c

+++ xorg-server/hw/xwin/wintrayicon.c
@@ -193,7 +193,7 @@ winHandleIconMessage(HWND hwnd, UINT message,
             RemoveMenu(hmenuTray, ID_APP_HIDE_ROOT, MF_BYCOMMAND);
         }
 
-        SetupRootMenu((unsigned long) hmenuTray);
+        SetupRootMenu(hmenuTray);
 
         /*
          * NOTE: This three-step procedure is required for
